# Make RNN faster by transposing weights