--- a/scripts/dump_rnn.py
+++ b/scripts/dump_rnn.py
@@ -12,7 +12,7 @@ from keras import backend as K
 import numpy as np
 
 def printVector(f, vector, name):
-    v = np.reshape(vector, (-1));
+    v = vector.T.reshape(-1)
     #print('static const float ', name, '[', len(v), '] = \n', file=f)
     f.write('static const opus_int16 {}[{}] = {{\n   '.format(name, len(v)))
     for i in range(0, len(v)):