Transpoing weights(changing memory layout) can speedup RNN about 2.5x.

↓ Before transposing
```
PS D:\work\opus\win32\VS2015\x64\Release> Measure-Command {.\opus_demo -e audio 16000 1 32000 rec2_2017_11_02_14_34_31.p
cm rec2_2017_11_02_14_34_31.opus}
libopus 1.3-dirty
Encoding 16000 Hz input at 32.000 kb/s in auto bandwidth with 320-sample frames.
average bitrate:              31.507 kb/s
maximum bitrate:              66.000 kb/s
active bitrate:               38.296 kb/s
bitrate standard deviation:    7.944 kb/s


Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 3
Milliseconds      : 517
Ticks             : 35172426
TotalDays         : 4.07088263888889E-05
TotalHours        : 0.000977011833333333
TotalMinutes      : 0.05862071
TotalSeconds      : 3.5172426
TotalMilliseconds : 3517.2426
```

↓ After transposing
```
PS D:\work\opus\win32\VS2015\x64\Release> Measure-Command {.\opus_demo -e audio 16000 1 32000 rec2_2017_11_02_14_34_31.p
cm rec2_2017_11_02_14_34_31.opus}
libopus 1.3-dirty
Encoding 16000 Hz input at 32.000 kb/s in auto bandwidth with 320-sample frames.
average bitrate:              31.507 kb/s
maximum bitrate:              66.000 kb/s
active bitrate:               38.296 kb/s
bitrate standard deviation:    7.944 kb/s


Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 3
Milliseconds      : 420
Ticks             : 34200928
TotalDays         : 3.95844074074074E-05
TotalHours        : 0.000950025777777778
TotalMinutes      : 0.0570015466666667
TotalSeconds      : 3.4200928
TotalMilliseconds : 3420.0928
```

It's about 2~3% speedup for opus encoding.

I also edited dumping scripts to transpose weights as needed. (Couldn't test it tho.)