@jmvalin 2018-11-03
There's actually a good reason (at least in theory) why the weights are the way the are. The current ordering should actually be faster once the code is properly vectorized. The transposed version in this patch is probably easier to vectorize for the compiler (hence why it's faster), but once properly vectorized, I think the current ordering will be faster because we can avoid the horizontal sums at the end (which the transposed version would need). Feel like doing a bit of vectorization instead?