@dofuuz 2018-11-04
Can you give some clue (keyword, link...) about optimization that you are going to use?
As I know, SIMD vectorization needs continuous memory access anyway.