@jmvalin 2018-11-04
Basically, consider a matrix-vector product like this one:

   for (i=0;i<N;i++)
   {
      /* Compute update gate. */
      float sum = layer->bias[i];
      for (j=0;j<M;j++)
         sum += layer->input_weights[j*stride + i]*input[j];
      output[i] = WEIGHTS_SCALE*sum;
   }

If you're using SSE, you'll want to unroll the outer loop (on i) by 4, so that each vector computes four outputs and each load from the weight matrix loads elements from 4 lines at a time. If you don't unroll the inner loop (on j), then you load one input[] element at a time and broadcast it to all vector elements before the multiply. However, if you want to go a little faster, then you can unroll the inner loop by 4 as well and just load once (4 values) and broadcast 4 times. You can also keep 2 or 4 separate partial sums so you don't get hit by the add latency. Get the general idea?